export speed module to be reusable into leaderboard
